added featureload files

+++ bottleneck.py
@@ -4,14 +4,9 @@ from keras.applications.vgg16 import preprocess_input
 import numpy as np
 from os import listdir
 from os.path import isdir, join, isfile
-import glob
+import datetime
 
 
-#model = ResNet50(include_top=False, weights='imagenet', input_tensor=None)
-#print("loaded Resnet")
-
-train_data_dir = '../pilz-scrapper/target/train'
-validation_data_dir = '../pilz-scrapper/target/test'
 
 def load_image(path="7.jpg"):
     img = image.load_img(path, target_size=(224, 224))
@@ -24,14 +19,17 @@ def load_image_folder(path):
     arrays = []
     image_paths = [join(path, fe) for fe in listdir(path) if isfile(join(path, fe))]
     for image_path in image_paths:
-        if not image_path.endswith(".jpg"):
+        if image_path.endswith(".npy"):
             continue
         array = load_image(image_path)
         arrays.append(array)
-
+    if len(arrays) == 0:
+        print(path)
     return np.concatenate(tuple(arrays))
 
 def save_bottleneck(categorie_path, model):
+    if isfile(categorie_path + "/features.npy"):
+        return
     x = load_image_folder(categorie_path)
     features = model.predict(x)
     np.save(categorie_path + "/features", features)
@@ -44,7 +42,13 @@ def save_bottlenecks(path):
     for categorie in onlydir:
         save_bottleneck(categorie, model)
         i+=1
-        print(str(i) + "/" + str(len(onlydir)))
+        print(str(datetime.datetime.now()) + ": " + str(i) + "/" + str(len(onlydir)))
 
 
 
+
+train_data_dir = '../pilz-scrapper/target/train'
+validation_data_dir = '../pilz-scrapper/target/test'
+
+save_bottlenecks(validation_data_dir)
+save_bottlenecks(train_data_dir)
